stack/requirements.txt
aws-cdk.core==1.102.0
aws-cdk.aws-lambda==1.102.0
aws-cdk.aws_ecr==1.102.0
aws-cdk.aws_iam==1.102.0
aws-cdk.aws_apigatewayv2==1.102.0
aws-cdk.aws_efs==1.102.0
aws-cdk.aws_ec2==1.102.0